% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brio-package.R
\docType{package}
\name{brio-package}
\alias{brio}
\alias{brio-package}
\title{brio: Basic R Input Output}
\description{
Functions to handle basic input output, these functions always read and write UTF-8 (8-bit Unicode Transformation Format) files and provide more explicit control over line endings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://brio.r-lib.org}
  \item \url{https://github.com/r-lib/brio}
  \item Report bugs at \url{https://github.com/r-lib/brio/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Authors:
\itemize{
  \item Jim Hester (\href{https://orcid.org/0000-0002-2739-7082}{ORCID})
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
