% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{brio}
\alias{brio}
\title{Brio Application}
\usage{
brio(...)
}
\arguments{
\item{...}{Optional parameters to be passed to \code{\link[shiny]{runApp}}.}
}
\value{
This function launches the app and returns nothing.
}
\description{
\code{brio} launches the `brio` shiny application.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
