% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_line_endings.R
\name{file_line_endings}
\alias{file_line_endings}
\title{Retrieve the type of line endings used by a file}
\usage{
file_line_endings(path)
}
\arguments{
\item{path}{A character string giving the file path.}
}
\value{
The line endings used, one of
\itemize{
\item '\\n' - if the file uses Unix line endings
\item '\\r\\n' - if the file uses Windows line endings
\item NA - if it cannot be determined
}
}
\description{
Retrieve the type of line endings used by a file
}
