% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_file.R
\name{write_file}
\alias{write_file}
\title{Write data to a file}
\usage{
write_file(text, path)
}
\arguments{
\item{text}{A character vector of length 1 with data to write.}

\item{path}{A character string giving the file path to write to.}
}
\value{
The UTF-8 encoded input text (invisibly).
}
\description{
This function differs from \code{\link[=write_lines]{write_lines()}} in that it writes the data in
\code{text} directly, without any checking or adding any newlines.
}
\examples{
tf <- tempfile()

write_file("some data\n", tf)

unlink(tf)
}
