% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_file_raw.R
\name{write_file_raw}
\alias{write_file_raw}
\title{Write data to a file}
\usage{
write_file_raw(raw, path)
}
\arguments{
\item{raw}{A raw vector with data to write.}

\item{path}{A character string giving the file path to write to.}
}
\description{
This function differs from \code{\link[=write_lines]{write_lines()}} in that it writes the data in
\code{text} directly, without any checking or adding any newlines.
}
\examples{
tf <- tempfile()

write_file_raw(as.raw(c(0x66, 0x6f, 0x6f, 0x0, 0x62, 0x61, 0x72)), tf)

unlink(tf)
}
