% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_lines.R
\name{writeLines}
\alias{writeLines}
\title{Write lines to a file}
\usage{
writeLines(text, con, sep = "\\n", useBytes)
}
\description{
This is a drop in replacement for \code{\link[base:writeLines]{base::writeLines()}} with restricted
functionality. Compared to \code{\link[base:writeLines]{base::writeLines()}} it:
\itemize{
\item Only works with file paths, not connections.
\item Uses \code{\link[=enc2utf8]{enc2utf8()}} to convert \code{text()} to UTF-8 before writing.
\item Uses \code{sep} uncoditionally as the line ending, regardless of platform.
\item The \code{useBytes} argument is ignored, with a warning.
}
}
