% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_lines.R
\name{write_lines}
\alias{write_lines}
\title{Write lines to a file}
\usage{
write_lines(text, path, eol = "\\n")
}
\arguments{
\item{text}{A character vector to write}

\item{path}{A character string giving the file path to write to.}

\item{eol}{The end of line characters to use between lines.}
}
\value{
The UTF-8 encoded input text (invisibly).
}
\description{
The text is converted to UTF-8 encoding before writing.
}
\examples{
tf <- tempfile()

write_lines(rownames(mtcars), tf)

unlink(tf)
}
